+++ src/battle-update.js
@@ -110,6 +110,7 @@ export function resetBattleVars() {
   battleSt.currentAllyAttacker = -1; battleSt.allyTargetIndex = -1; battleSt.allyHitResult = null; battleSt.allyHitIsLeft = false;
   battleSt.allyShakeTimer = {}; battleSt.enemyTargetAllyIdx = -1; battleSt.forcedEnemyTarget = null;
   battleSt.allyMagicCasterIdx = -1; battleSt.allyMagicTargetIdx = -1; battleSt.allyMagicSpellId = 0;
+  battleSt.allySummonEffect = null;
   battleSt.allyMagicHealAmount = 0; battleSt.allyMagicDamageRoll = 0;
   battleSt.allyMagicEffectApplied = false; battleSt.allyMagicSfxPlayed = false; battleSt.allyMagicTargetType = 'player';
   hudSt.playerDeathTimer = null;